fix(duration): widen Duration.days from short to int — 89y silent-truncation bug

+++ src/main/java/io/brackit/query/atomic/YMD.java
@@ -313,7 +313,7 @@ public class YMD extends AbstractDuration {
   }
 
   @Override
-  public short getDays() {
+  public int getDays() {
     return 0;
   }
